Generate random options for a question
