+++ src/pages/quiz/QnAnsContainer.js
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from "react";
 import Wrapper from "../../components/wrapper";
 
 import words from "../../assets/data/words.json";
+import { randomUnique, questionObject } from "../../utils/QuizGenerator";
 
 export default function QnAnsContainer({ formInput }) {
   const [questions, setQuestions] = useState({
@@ -28,58 +29,32 @@ export default function QnAnsContainer({ formInput }) {
      Get Random indexes within a range
     */
     const indexes = randomUnique(value, ...range);
-    const { word, meaning } = words[indexes[0]];
-    const question = {
-      word,
-      answer: null,
-      options: [],
-      myChoice: null,
-      status: true,
-    };
+    const question = questionObject(indexes[0]);
 
-    /*
-     Get 3 Random indexes within a range for the option field
-    */
-    const option_indexes = randomUnique(3, ...range, indexes[0]);
-    const options = [];
-    option_indexes.forEach((index) => {
-      options.push(words[index].meaning);
-    });
-
-    /*
-     Get a Random index for the answer for the options list
-    */
-    const answer_index = randomUnique(1, ...[0, 4])[0];
-    /*
-     Insert answer in the random position generated
-    */
-    options.splice(answer_index, 0, meaning);
-    /*
-     Update the state object
-    */
-    question.options = [...options];
-    question.answer = answer_index;
+    indexes.shift();
 
     setQuestions({ indexes, question: [{ ...question }] });
   }, [formInput]);
 
-  const randomUnique = (count, min, max, ignoredIndex = null) => {
-    let nums = new Set();
-    while (nums.size < count) {
-      const index = Math.floor(
-        Math.floor(Math.random() * (max - min + 1) + min)
-      );
-      if (ignoredIndex === index) continue;
-      nums.add(index);
-    }
-    return [...nums];
-  };
+  function nextQuestion() {
+    const new_questions = { ...questions };
+    if (new_questions.indexes.length === 0) return alert("DONE");
+
+    /*
+     * Generate Question object
+     */
+    const question = questionObject(new_questions.indexes[0]);
+
+    new_questions.question.push(question);
+    new_questions.indexes.shift();
+    setQuestions({ ...new_questions });
+  }
 
   console.log(questions);
 
   return (
     <Wrapper contain spaceAround grow>
-      Hello
+      <button onClick={nextQuestion}>Click</button>
     </Wrapper>
   );
 }
